+++ dialect.go
@@ -62,7 +62,11 @@ func wrapPageSQL(dialect string, sqlstr *string, page *Page) error {
 		sqlbuilder.WriteString(strconv.Itoa(page.PageSize))
 		sqlbuilder.WriteString(" OFFSET ")
 		sqlbuilder.WriteString(strconv.Itoa(page.PageSize * (page.PageNo - 1)))
-	case "mssql", "oracle": //sqlserver 2012+,oracle 12c+  查询语句需要有order by
+	case "mssql", "oracle": //sqlserver 2012+,oracle 12c+
+		locOrderBy := findOrderByIndex(sqlstr)
+		if len(locOrderBy) <= 0 { //如果没有 order by,增加默认的排序
+			sqlbuilder.WriteString(" ORDER BY (SELECT NULL) ")
+		}
 		sqlbuilder.WriteString(" OFFSET ")
 		sqlbuilder.WriteString(strconv.Itoa(page.PageSize * (page.PageNo - 1)))
 		sqlbuilder.WriteString(" ROWS FETCH NEXT ")